Fix replayers' entrypoint script
we do want the main swh command to receive SIGTERM from docker so it
does gracefully disconnect the kafka consumer, thus properly leave the
consumer group.

This prevents from having to wait for the kafka server to figure that a
client actually left the consumer group before rebalancing the
partitions.

TL;DR: allows to make a restart/rescaling of the replayers much
faster.