Add support for an cassandra-based storage option
This comes as a dedicated compose file (trying to factorize it is a
PITA, neither 'docker compose config' nor multi compose file approach
prove to be working good enough for this).

Tests are executed both using the default postgresql based
storage (test_mirror.py) and the new cassandra based one
(test_mirror_cassandra.py) by default. The whole test session is 20 to
30mn long depending on the hardware.

The cassandra deployment provided in this compose file is by no mean
intended for anything close to a production-like deployment, it's barely
an example of an easy to test cassandra based deployment of a mirror,
nothing more.